move UserService to auth.user
